+++ NEWS.md
@@ -17,7 +17,7 @@ CHANGES
   dimnames. (#699 @elinw)
 
 * `distribution_mode()` now displays an informative message when mode frequencies
-  are tied (#702).
+  are tied, and tied values are stored in output attributes (#702).
 
 # datawizard 1.3.1